Add files via upload
This is an example of the 6 files I used to run the code. I altered the localfit_v2.py code to localfit_1000files to make it run however many copies that i want. The 'CreateMeanSTDCSV' produces a mean and standard deviation file that are then both used by the 'ComparePlotFPlot graph.